+++ protocol/grpc/internal/README.md
@@ -6,3 +6,8 @@ This internal package is for test. So don't use internal in production.
 
 - protoc
 - protoc-gen-go: github.com/golang/protobuf/protoc-gen-go@v1.3.0
+
+## usage
+
+- [helloworld](./hellowrld): test unary process
+- [routeguide](./routeguide): test stream process
