+++ protocol/grpc/internal/helloworld/helloworld.proto
@@ -21,7 +21,7 @@ option java_package = "io.grpc.examples.helloworld";
 option java_outer_classname = "HelloWorldProto";
 option objc_class_prefix = "HLW";
 
-package internal;
+package helloworld;
 
 // The greeting service definition.
 service Greeter {
